% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_data.R
\name{check_meta_data}
\alias{check_meta_data}
\title{Check overview data for potential errors}
\usage{
check_meta_data(overview_excel)
}
\arguments{
\item{overview_excel}{Excel overview data}
}
\description{
Check overview data for potential errors
}
