% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{concat_excel}
\alias{concat_excel}
\title{Concatenating the excel files}
\usage{
concat_excel(desc_path, desc_continent, desc_startpage, desc_endpage)
}
\arguments{
\item{desc_path}{The list pointing to the path of the excel files}

\item{desc_continent}{The continent of the excel file}

\item{desc_startpage}{The startpage of the excel file}

\item{desc_endpage}{The endpage of th excel file}
}
\description{
Concatenating the excel files
}
