% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{concat_sheets}
\alias{concat_sheets}
\title{Concatenating Excel-files}
\usage{
concat_sheets()
}
\description{
This function concatenates the individual Excel-files to one
R-file
}
\keyword{internal}
