% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correcting_region_ids.R
\name{correcting_region_ids}
\alias{correcting_region_ids}
\title{Correcting Region Data}
\usage{
correcting_region_ids()
}
\value{
header_data_corrected
}
\description{
Removing invalid IDs, meaning regent IDs that have been entered
wrong and include typos.
}
