% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correction_required}
\alias{correction_required}
\title{Manual Correcting the Headers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1855 rows and 8 columns.
}
\usage{
correction_required
}
\description{
Headers that are either missing or unspecified and require
a manual correction of the excel sheets that created them.
}
\keyword{datasets}
