% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{create_final_sheet}
\alias{create_final_sheet}
\title{Creating the final state of the excel files in R form}
\usage{
create_final_sheet()
}
\value{

}
\description{
Creating the final state of the excel files in R form
}
