% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{desc_file}
\alias{desc_file}
\title{This filee provides an overview of the individual excel files}
\usage{
desc_file()
}
\description{
This filee provides an overview of the individual excel files
}
