% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{excel_count}
\alias{excel_count}
\title{Counting the number of excel sheets currently residing inside the explorer}
\usage{
excel_count(x)
}
\description{
Counting the number of excel sheets currently residing inside the explorer
}
