% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_base_dataset.R
\name{gen_base_dataset}
\alias{gen_base_dataset}
\title{Creating Base dataframe}
\usage{
gen_base_dataset()
}
\value{
input_data
}
\description{
Dataframe that will be used as the "building foundation" for all
the following data cleaning functions.
}
