% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{gen_dataset}
\alias{gen_dataset}
\title{Creating Base dataframe}
\usage{
gen_dataset()
}
\value{
input_data
}
\description{
Dataframe that will be used as the "building foundation" for all
the following data cleaning functions.
}
