% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_estimating_period.R
\name{gen_estimation}
\alias{gen_estimation}
\title{Estimate Period}
\usage{
gen_estimation()
}
\description{
Function to estimate the period and related variables like
when the reign of the regent started, when it ended etc.
}
