% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_group_id.R
\name{gen_group_id}
\alias{gen_group_id}
\title{Generate Group Identifier}
\usage{
gen_group_id()
}
\value{
group_id_dataset
}
\description{
The group identifier is the most important part when assigning
ruler to administrative regions as it links the individuals to the
administrative regions
}
