% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{gen_header}
\alias{gen_header}
\title{Generate header}
\usage{
gen_header(sheet)
}
\arguments{
\item{sheet}{Sheet generated by function concat_sheets}
}
\description{
The headers contain useful information i.e. the region and the
country of the ruler. To extract these information, use is made of the fact
that the left contains the region and the right header the country.
}
