% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meta_data.R
\name{gen_meta_data}
\alias{gen_meta_data}
\title{Generate Meta Data}
\usage{
gen_meta_data(input_path)
}
\arguments{
\item{input_path}{Path to the excel files}
}
\value{
Data frame with a single row and column names and values
corresponding to the meta data found in the respective excel-file.
}
\description{
Generate Meta Data
}
