% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{gen_overview}
\alias{gen_overview}
\title{Generating overview}
\usage{
gen_overview(path_input = "D:/km/Truhart")
}
\arguments{
\item{path_input}{Path to the directory where the Truhart files are saved}
}
\description{
This function generates an overview of the available
files
}
