% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_reign_summary.R
\name{gen_reign_summary}
\alias{gen_reign_summary}
\title{Reign Summary}
\usage{
gen_reign_summary()
}
\description{
This function generates a summary about the reign of the regents
including the length of the reign, the start and end year (if given) and
the century and decade the ruler ruled in.
}
\author{
Tim Holzapfel
}
