% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unique_index_range.R
\name{gen_unique_index_range}
\alias{gen_unique_index_range}
\title{Unique Index range}
\usage{
gen_unique_index_range()
}
\description{
Function to determine and return the unique index level of
all administrative ares
}
