% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{generate_sheets}
\alias{generate_sheets}
\title{Concatenating Excel-files}
\usage{
generate_sheets(path_input = "D:/km/Truhart")
}
\arguments{
\item{path_input}{The path to the Truhart Excel-files}
}
\description{
This function concatenates the individual Excel-files to one
R-file
}
