% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_headers}
\alias{missing_headers}
\title{Manual Correcting the Headers}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 122 rows and 7 columns.
}
\usage{
missing_headers
}
\description{
Headers that are either missing or unspecified and require
a manual correction of the excel sheets that created them.
}
\keyword{datasets}
