% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocrProject-package.R
\docType{package}
\name{ocrProject-package}
\alias{ocrProject}
\alias{ocrProject-package}
\title{ocrProject: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tim Holzapfel \email{TimHolzapfel@outlook.com}

}
\keyword{internal}
