% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{open_stata}
\alias{open_stata}
\title{Open File in Stata}
\usage{
open_stata(file)
}
\arguments{
\item{file}{R file that is to be opened in Stata}
}
\value{
Stata file
}
\description{
The Stata data viewer is clearly superior to the R data viewer
at the moment. Thus it makes sense to utilize the capabilities of both
programs to achieve maximum results.
}
