% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_files.R
\name{read_excel_files}
\alias{read_excel_files}
\title{Creating Dataframes from Excel Files}
\usage{
read_excel_files(
  input_path,
  input_continent,
  input_region,
  input_start_page,
  input_end_page
)
}
\arguments{
\item{input_path}{Path to the excel files}

\item{input_continent}{Name of the continent the Excel file refers to}

\item{input_region}{Name of the continent region the Excel file refers to}

\item{input_start_page}{Starting page of the Excel file}

\item{input_end_page}{Ending page of the Excel file}
}
\value{
Excel sheet in data frame format
}
\description{
Creating Excel-sheets from the available files
}
