% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regents_of_nations}
\alias{regents_of_nations}
\title{Final dataset produced by this package}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 76209 rows and 34 columns.
}
\usage{
regents_of_nations
}
\description{
Final Dataset ready for use in statistical analysis
}
\keyword{datasets}
