% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removing_invalid_ids.R
\name{removing_invalid_ids}
\alias{removing_invalid_ids}
\title{Correcting Region Data}
\usage{
removing_invalid_ids()
}
\value{
header_data_corrected
}
\description{
Removing invalid IDs, meaning regent IDs that have been entered
wrong and include typos.
}
