% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_empty.R
\name{replace_empty}
\alias{replace_empty}
\title{Replace Empty with Missing}
\usage{
replace_empty(data_file)
}
\arguments{
\item{data_file}{Dataframe to be modified}
}
\description{
Function to replace empty strings with NA_character_.
}
