% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman2numeric.R
\name{roman2numeric}
\alias{roman2numeric}
\title{roman2numeric}
\usage{
roman2numeric(x)
}
\arguments{
\item{x}{roman number}
}
\description{
roman2numeric
}
\note{
This code was taken directly from the R base package utils and needed
to be extracted because it utils did not export it.
}
\author{
R Core Team
}
