% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman_to_alpha.R
\name{roman_to_alpha}
\alias{roman_to_alpha}
\title{Roman To Alpha}
\usage{
roman_to_alpha(x)
}
\arguments{
\item{x}{Roman number}
}
\value{
Abbreviated name of the month
}
\description{
This function takes a roman numeral as argument, replaces it
with an Arabic number and finally returns the month the roman numeral
represented based on the Arabic number
}
