% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruler_subset.R
\name{ruler_subset}
\alias{ruler_subset}
\title{Data subset only containing ruler}
\usage{
ruler_subset(data)
}
\arguments{
\item{data}{Data set containing at least an id column and the names of rulers
to subset for}
}
\value{
Same data set as before but with fewer observations.
}
\description{
Subsetting the data based on the id and selecting only
observation that are composed of only three digits. These indicate the
presence of an actual ruler as opposed to rows describing attributes of the
aforementioned ruler.
}
