% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_squish.R
\name{string_squish}
\alias{string_squish}
\title{String Squish}
\usage{
string_squish(dataframe_tb)
}
\arguments{
\item{dataframe_tb}{dataframe that should be squished}
}
\description{
Removing leading, trailing and duplicate spaces from all string
characters of a dataframe
}
\note{
One thing one has to be careful about is that "str_squish" will return
variables it has been applied upon as type character. This function
therefore only selects columns of type character to keep non-string columns
in the original type.
}
