% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_concat_files_md.R
\name{concat_excel}
\alias{concat_excel}
\title{concat_excel}
\usage{
concat_excel(desc_path, desc_continent, desc_startpage, desc_endpage)
}
\arguments{
\item{path_input}{The list pointing to the path of the excel files that are to be merged}
}
\value{

}
\description{
concat_excel
}
