% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_header.R
\name{gen_header}
\alias{gen_header}
\title{Generate header}
\usage{
gen_header(dev_mode = FALSE)
}
\arguments{
\item{dev_mode}{When set to TRUE stops the execution early and returns the
intermediate dataset required for testing}
}
\description{
The headers contain useful information i.e. the region and the
country of the ruler. To extract these information, use is made of the fact
that the left contains the region and the right header the country.
}
\author{
Tim Holzapfel
}
