% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_overview.R
\name{gen_overview}
\alias{gen_overview}
\title{Generating overview}
\usage{
gen_overview()
}
\value{
overview_data
}
\description{
This function generates an overview of the available files
}
\author{
Tim Holzapfel
}
