% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{gen_overview}
\alias{gen_overview}
\title{Generating overview}
\usage{
gen_overview(input_path = "D:/km/Truhart")
}
\arguments{
\item{input_path}{Path to the directory where the Truhart files are saved}
}
\value{
overview_data
}
\description{
This function generates an overview of the available files
}
\author{
Tim Holzapfel
}
