% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocrProject-package.R
\docType{package}
\name{ocrProject-package}
\alias{ocrProject}
\alias{ocrProject-package}
\title{ocrProject: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
