% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocrProject-package.R
\docType{package}
\name{ocrProject-package}
\alias{ocrProject}
\alias{ocrProject-package}
\title{ocrProject: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Tim-Lukas-H/ocrProject}
  \item Report bugs at \url{https://github.com/Tim-Lukas-H/ocrProject/issues}
}

}
\author{
\strong{Maintainer}: Tim Holzapfel \email{timholzapfel@outlook.com} (\href{https://orcid.org/0000-0002-2019-520X}{ORCID})

}
\keyword{internal}
