% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocrProject-package.R
\docType{package}
\name{ocrProject-package}
\alias{ocrProject}
\alias{ocrProject-package}
\title{ocrProject: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tim Holzapfel \email{timholzapfel@outlook.com} (\href{https://orcid.org/0000-0002-2019-520X}{ORCID})

}
\keyword{internal}
