% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{read_excel_files}
\alias{read_excel_files}
\title{Creating Dataframes from Excel Files}
\usage{
read_excel_files(
  input_path,
  input_continent = NA,
  input_startpage = NA,
  input_endpage = NA,
  input_pdf_file = NA
)
}
\arguments{
\item{input_path}{Path to the directory where the Truhart files are saved}

\item{input_continent}{Name of the continent the Excel file refers to}

\item{input_startpage}{Starting page of the Excel file}

\item{input_endpage}{Ending page of the Excel file}

\item{input_pdf_file}{Pdf file corresponding to the excel file}
}
\value{
Excel sheet in data frame format
}
\description{
Creating Excel-sheets from the available files
}
